+++ src/main/java/algo/sarsa/SarsaAlgo.java
@@ -9,14 +9,14 @@ public class SarsaAlgo implements BaseAlgo {
     /** Root node of the last call to {@code nextMove}. */
     private SarsaNode root;
     /** The move that the agent last chose. */
-    private int move;
+    private byte move;
     /** Probability of making a random move. */
     private final double p = 0.1;
     /** Number of training epochs. */
     private final int epochs = 500000;
 
     @Override
-    public int nextMove(Board board) {
+    public byte nextMove(Board board) {
         this.setupRoot(board);
         for (int i = 0; i < this.epochs; i++) {
             if (i % 10000 == 0) {
@@ -39,7 +39,7 @@ public class SarsaAlgo implements BaseAlgo {
     }
 
     @Override
-    public int nextMoveWithTime(Board board, int time) {
+    public byte nextMoveWithTime(Board board, int time) {
         this.setupRoot(board);
         long endTime = System.currentTimeMillis() + time;
         while (System.currentTimeMillis() < endTime) {
@@ -61,7 +61,7 @@ public class SarsaAlgo implements BaseAlgo {
     }
 
     @Override
-    public List<Integer> getMovePredictions() {
+    public List<Byte> getMovePredictions() {
         assert this.root != null;
         return this.root.bestMoveSequence();
     }
